refactor: process step 파일, 상수 분리
