+++ src/components/Dashboard/ProcessSection.tsx
@@ -1,8 +1,7 @@
 import { Flex, Heading, Text } from '@/components';
-import { MAX_RESPONSIVE_WIDTH } from '@/constants/system';
+import ProcessStep from '@/components/Dashboard/ProcessStep';
 import { boxShadow } from '@/styles/common';
-import { getOpacityColor } from '@/utils/getOpacityColor';
-import { styled, useMediaQuery, useTheme } from '@mui/material';
+import { styled, useTheme } from '@mui/material';
 
 const ProcessSection = () => {
   const theme = useTheme();
@@ -34,45 +33,6 @@ const ProcessSection = () => {
 
 export default ProcessSection;
 
-const processStep = [
-  '마일리지 시스템에 신청',
-  '신분증 및 통장 사본 제출',
-  '개인정보 이용 동의서 서명 완료',
-];
-
-const ProcessStep = () => {
-  const isMobile = useMediaQuery(MAX_RESPONSIVE_WIDTH);
-  return (
-    <>
-      {processStep.map((step, index) => (
-        <>
-          <S.StepBox
-            key={`step-${step}`}
-            justify="center"
-            align="center"
-            isMobile={isMobile}
-          >
-            <Heading
-              as="h3"
-              style={{
-                textAlign: 'center',
-              }}
-            >
-              {step}
-            </Heading>
-          </S.StepBox>
-          {index + 1 !== processStep.length &&
-            (isMobile ? (
-              <S.DownArrowBox key={`down-arrow-${step}`} />
-            ) : (
-              <S.RightArrowBox key={`right-arrow-${step}`} />
-            ))}
-        </>
-      ))}
-    </>
-  );
-};
-
 const S = {
   Container: styled(Flex.Row)`
     background-color: ${({ theme }) => theme.palette.white};
@@ -80,31 +40,5 @@ const S = {
     padding: 1rem;
 
     ${boxShadow}
-  `,
-  StepBox: styled(Flex.Row)<{ isMobile: boolean }>`
-    background-color: ${({ theme }) => theme.palette.primary.light};
-    border: 2px solid
-      ${({ theme }) => getOpacityColor(theme.palette.primary.main, 0.5)};
-    border-radius: 1rem;
-    color: ${({ theme }) => theme.palette.primary.main};
-    height: fit-content;
-    padding: ${({ isMobile }) => (isMobile ? '1rem' : '1.5rem')};
-    width: ${({ isMobile }) => (isMobile ? '100%' : 'fit-content')};
-  `,
-  DownArrowBox: styled('div')`
-    border-left: 20px solid transparent;
-    border-right: 20px solid transparent;
-    border-top: 25px solid
-      ${({ theme }) => getOpacityColor(theme.palette.primary.main, 0.7)};
-    height: 0;
-    width: 0;
-  `,
-  RightArrowBox: styled('div')`
-    border-bottom: 20px solid transparent;
-    border-left: 25px solid
-      ${({ theme }) => getOpacityColor(theme.palette.primary.main, 0.7)};
-    border-top: 20px solid transparent;
-    height: 0;
-    width: 0;
   `,
 };
