+++ src/components/Dashboard/index.ts
@@ -3,5 +3,6 @@ export { default as FAQSection } from './FAQSection';
 export { default as LineChartSection } from './LineChartSection';
 export { default as MileageCountSection } from './MileageCountSection';
 export { default as ProcessSection } from './ProcessSection';
+export { default as ProcessStep } from './ProcessStep';
 export { default as RadarChartSection } from './RadarChartSection';
 export { default as ScholarshipDurationSection } from './ScholarshipDurationSection';
