+++ src/constants/processStep.ts
@@ -1,0 +1,5 @@
+export const processStep = [
+  '마일리지 시스템에 신청',
+  '신분증 및 통장 사본 제출',
+  '개인정보 이용 동의서 서명 완료',
+];
